% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suptCriticalValue.R
\name{suptCriticalValue}
\alias{suptCriticalValue}
\title{Create sup t critical value}
\usage{
suptCriticalValue(
  vcov_matrix,
  num_sim = 1000,
  conf_level = 0.95,
  seed = 192837
)
}
\arguments{
\item{vcov_matrix}{Variance-covariance matrix}

\item{num_sim}{Number of bootstrap simulations. Default is 1000.}

\item{conf_level}{Confidence level. Default is 0.95.}

\item{seed}{Random seed. Default is 192837.}
}
\value{
Numeric. Sup t critical value. Multiple by the standard error to get the confidence interval length.
}
\description{
Create sup t critical value
}
\examples{
library(MASS)
data(Cars93, package = "MASS")
conf_level  <- 0.95
model       <- lm(Price ~ Weight + Wheelbase + Cylinders, data=Cars93)

beta        <- matrix(coef(model))
vcov_matrix <- vcov(model)
std_error   <- sqrt(diag(vcov_matrix))

pw_crit     <- qt(1 - ((1 - conf_level) / 2), model$df.residual) 
supt_crit   <- suptCriticalValue(vcov_matrix = vcov_matrix)

pw_ci_lb    <- beta - pw_crit * std_error 
pw_ci_ub    <- beta + pw_crit * std_error 
supt_ci_lb  <- beta - supt_crit * std_error 
supt_ci_ub  <- beta + supt_crit * std_error 

print("POINTWISE 95 PERCENT CONFIDENCE INTERVAL: ")
print(paste0(
  "[", pw_ci_lb, ", ", pw_ci_ub, "]"
))

print("SIMULTANEOUS SUP-T 95 PERCENT CONFIDENCE INTERVAL: ")
print(paste0(
  "[", supt_ci_lb, ", ", supt_ci_ub, "]"
))

}
